# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StreamPay API",
    "version": "2.0.0",
    "description": "REST API for the StreamPay payment-streaming platform on Stellar.\n\nAll error responses use the canonical envelope:\n```json\n{ \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"Human-readable detail.\", \"request_id\": \"req_...\" } }\n```\n\nv1 stream paths (`/api/streams`) are deprecated and will be removed. Use `/api/v2/streams`."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Obtain a token via POST /api/auth/wallet, then pass it as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "request_id"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code.",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "Human-readable description safe to display to clients.",
                "example": "The requested stream does not exist."
              },
              "request_id": {
                "type": "string",
                "description": "Opaque identifier for this request, useful for log correlation.",
                "example": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "WalletChallenge": {
        "type": "object",
        "required": ["challenge", "expires_at"],
        "properties": {
          "challenge": {
            "type": "string",
            "description": "One-time nonce the client must sign with their Stellar private key.",
            "example": "streampay_auth_1716000000000_x7k2m"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Challenge expiry (5 minutes from issuance).",
            "example": "2024-01-15T10:05:00.000Z"
          }
        }
      },
      "WalletToken": {
        "type": "object",
        "required": ["token", "expires_at"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque bearer token. Pass as `Authorization: Bearer <token>`.",
            "example": "tok_R0FCQzEyMzQ1Njc4OTA"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Token expiry (24 hours from issuance).",
            "example": "2024-01-16T10:00:00.000Z"
          }
        }
      },
...
      "ActivityEvent": {
        "type": "object",
        "required": ["id", "type", "timestamp", "description"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique event identifier.",
            "example": "evt_a1b2c3d4"
          },
          "type": {
            "type": "string",
            "description": "Event type classification.",
            "example": "stream.created"
          },
          "streamId": {
            "type": "string",
            "description": "Associated stream identifier, if applicable.",
            "example": "stream-abc12345"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 timestamp of the event.",
            "example": "2026-04-28T09:00:00Z"
          },
          "description": {
            "type": "string",
            "description": "Human-readable event description.",
            "example": "Stream 'Design Retainer' created and set to draft."
          }
        }
      },
      "PaginatedMeta": {
        "type": "object",
        "required": ["hasNext", "nextCursor", "total"],
        "properties": {
          "hasNext": {
            "type": "boolean",
            "description": "Whether additional results are available."
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Opaque base64-encoded cursor for the next page. Null when hasNext is false.",
            "example": "MjAyNi0wNC0yN1QyMDowMDowMFp8ZXZ0LTI="
          },
          "total": {
            "type": "integer",
            "description": "Total number of items matching the filter criteria.",
            "example": 42
          }
        }
      },
      "ActivityResponse": {
        "type": "object",
        "required": ["data", "meta", "links"],
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ActivityEvent" }
          },
          "meta": { "$ref": "#/components/schemas/PaginatedMeta" },
          "links": {
            "type": "object",
            "required": ["self"],
            "properties": {
              "self": {
                "type": "string",
                "description": "Canonical URL for this request.",
                "example": "/api/activity?limit=20"
              }
            }
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad request — invalid input.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request body must include 'recipient' and 'rate'.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized — missing or invalid bearer token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Bearer token required.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "An unexpected error occurred.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/activity": {
      "get": {
        "operationId": "listActivity",
        "summary": "List activity events",
        "description": "Returns a paginated, cursor-based list of activity events. Events are sorted by (timestamp desc, id desc) for stable ordering. Supports filtering by streamId and event type.",
        "tags": ["Activity"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of events to return (1-100, default 20).",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque base64-encoded composite cursor (timestamp|id) for cursor-based pagination. Obtained from meta.nextCursor of the previous response.",
            "schema": { "type": "string" }
          },
          {
            "name": "streamId",
            "in": "query",
            "description": "Filter events by associated stream identifier.",
            "schema": { "type": "string" }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter events by type (e.g. stream.created, wallet.connected).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of activity events.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityResponse" }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity — malformed cursor.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            }
          },
          "429": {
            "description": "Too many requests — rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/wallet": {
      "get": {
        "operationId": "getWalletChallenge",
        "summary": "Issue a wallet authentication challenge",
        "description": "Returns a one-time nonce the client must sign with their Stellar private key, then submit to `POST /api/auth/wallet`.",
        "security": [],
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "description": "Stellar public key (G…, 56 characters).",
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z2-7]{55}$",
              "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge issued.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WalletChallenge" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/400" },
          "500": { "$ref": "#/components/responses/500" }
        }
      },
      "post": {
        "operationId": "verifyWalletSignature",
        "summary": "Verify a signed challenge and issue a bearer token",
        "security": [],
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["address", "challenge", "signature"],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Stellar public key.",
                    "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
                  },
                  "challenge": {
                    "type": "string",
                    "description": "The nonce returned by GET /api/auth/wallet.",
                    "example": "streampay_auth_1716000000000_x7k2m"
                  },
                  "signature": {
                    "type": "string",
                    "description": "Base64-encoded Ed25519 signature of the challenge.",
                    "example": "dGVzdHNpZ25hdHVyZQ=="
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WalletToken" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/400" },
          "401": { "$ref": "#/components/responses/401" },
          "500": { "$ref": "#/components/responses/500" }
        }
      }
    },
...
  },
...
}